move decompose methods into their own file
